password_manager.py moved to domain
