+++ src/transcriber_service/domain/password_manager.py
